requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.2.0
django-oauth-toolkit==1.1.3
djangorestframework==3.10.1
-e git+https://github.com/tfranzel/djangorestframework-camel-case.git@bd556d38fa7382acadfe91d93d92d99c663248a9#egg=djangorestframework_camel_case
djangorestframework-recursive==0.1.2
drf-yasg==1.16
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
oauthlib==3.0.2
packaging==19.0
Pillow==6.1.0
pyparsing==2.4.0
pyrsistent==0.15.3
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
ruamel.yaml==0.15.100
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.25.3
user-agents==2.0